# internal/app/adapter/server/http/api/docs/swagger.yaml
basePath: /api
consumes:
- application/json
- application/x-www-form-urlencoded
definitions:
  example.ClientError:
    properties:
      errMsg:
        example: 客户端请求错误|参数校验错误
        type: string
      errNo:
        description: errNo 类型应为 int，string 仅为了表达多个错误码
        example: 10002|10003
        type: string
    type: object
  example.PermissionDenied:
    properties:
      errMsg:
        example: 暂无权限
        type: string
      errNo:
        example: 10005
        type: integer
    type: object
  example.ResourceNotFound:
    properties:
      errMsg:
        example: 资源不存在
        type: string
      errNo:
        example: 10006
        type: integer
    type: object
  example.ServerError:
    properties:
      errMsg:
        example: 服务器出错
        type: string
      errNo:
        example: 10001
        type: integer
    type: object
  example.Success:
    properties:
      data: {}
    type: object
  example.TooManyRequest:
    properties:
      errMsg:
        example: 请求过于频繁
        type: string
      errNo:
        example: 10007
        type: integer
    type: object
  example.Unauthorized:
    properties:
      errMsg:
        example: 未经授权
        type: string
      errNo:
        example: 10004
        type: integer
    type: object
...
  v1.SystemPermissionInfo:
    properties:
      desc:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      parentID:
        type: integer
    type: object
...
  v1.SystemRoleInfo:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  v1.SystemSessionLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  v1.SystemSessionLoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/v1.SystemSessionProfileResponse'
    type: object
  v1.SystemSessionProfileResponse:
    properties:
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/v1.SystemRoleInfo'
        type: array
      username:
        type: string
    type: object
  v1.SystemSessionRegisterRequest:
    properties:
      nickname:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  v1.SystemSessionRegisterResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/v1.SystemUserInfo'
    type: object
  v1.SystemSessionUpdatePasswordRequest:
    properties:
      oldPassword:
        type: string
      password:
        type: string
    type: object
  v1.SystemSessionUpdateProfileRequest:
    properties:
      nickname:
        type: string
    type: object
...
  v1.SystemUserInfo:
    properties:
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: API 接口文档
  title: API 接口文档
  version: 0.0.0
paths:
...
  /v1/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 请求体
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemSessionLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  $ref: '#/definitions/v1.SystemSessionLoginResponse'
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 登录
      tags:
      - 账号
  /v1/logout:
    delete:
      consumes:
      - text/plain
      description: 登出
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 登出
      tags:
      - 账号
  /v1/permissions:
    get:
      consumes:
      - text/plain
      description: 获取账号权限
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/v1.SystemPermissionInfo'
                  type: array
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 获取账号权限
      tags:
      - 账号
...
  /v1/profile:
    get:
      consumes:
      - text/plain
      description: 获取账号信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  $ref: '#/definitions/v1.SystemSessionProfileResponse'
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 获取账号信息
      tags:
      - 账号
    put:
      consumes:
      - application/json
      description: 更新账号信息
      parameters:
      - description: 请求体
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemSessionUpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 更新账号信息
      tags:
      - 账号
  /v1/profile/password:
    put:
      consumes:
      - application/json
      description: 修改当前登录用户密码。新密码 8～18 位，且同时包含数字、大写字母、小写字母和符号
      parameters:
      - description: 原密码与新密码
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemSessionUpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 修改当前登录用户密码
      tags:
      - 账号
  /v1/register:
    post:
      consumes:
      - application/json
      description: 注册
      parameters:
      - description: 请求体
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemSessionRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  $ref: '#/definitions/v1.SystemSessionRegisterResponse'
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 注册
      tags:
      - 账号
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"